web/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fonttools==4.28.1
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.2
Pillow==8.4.0
psycopg2 @ file:///opt/concourse/worker/volumes/live/7541fe58-2a45-4073-5d69-313e733fc729/volume/psycopg2_1612298592535/work
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tomli==1.2.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
